requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
jinja2==3.1.2
psutil==5.9.6
websockets==12.0
